add mixin for shared smooth scroll logic
